Implemented NextAuth Authentication
To be able to use the spotify services, a login process is required. This will communicate to the spotify api and use login credentials. An access token is required to access the playlists, songs and play the songs. (These expire every hour, thus refresh tokens required to create new access tokens)